# lib/api-zod/src/generated/types/openaiImageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TikTok Dropship AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { OpenaiImageInputSize } from "./openaiImageInputSize";

export interface OpenaiImageInput {
  prompt: string;
  size?: OpenaiImageInputSize;
}