# lib/api-zod/src/generated/types/supplierResearchInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TikTok Dropship AI Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface SupplierResearchInput {
  productName: string;
  productCategory: string;
  count?: number;
}